refactor: extract dispatcher class from store
