+++ packages/store/src/module.ts
@@ -3,6 +3,7 @@ import { NgModule, ModuleWithProviders, Optional, Inject } from '@angular/core';
 import { ROOT_STATE_TOKEN, FEATURE_STATE_TOKEN } from './symbols';
 import { StateFactory } from './state-factory';
 import { Actions, InternalActions } from './actions-stream';
+import { InternalDispatcher } from './dispatcher';
 import { Store } from './store';
 import { SelectFactory } from './select';
 import { StateStream } from './state-stream';
@@ -103,6 +104,7 @@ export class NgxsModule {
         StateFactory,
         Actions,
         InternalActions,
+        InternalDispatcher,
         Store,
         StateStream,
         SelectFactory,
